Zmiany dalsze

+++ src/bash/master.sh
@@ -1,10 +1,19 @@
 #!/bin/bash
 
+export FACTER_repopath="$1"
+export environment="$2"
+
 set -e
 set -x
 
-apt-get install -y puppetmaster ruby
-export FACTER_repopath="$1"
-environment="$2"
-puppet apply ${FACTER_repopath}/site/profile/manifests/master/provision-base.pp --environment $environment
-puppet apply ${FACTER_repopath}/site/profile/manifests/master/provision-finish.pp --environment $environment
+${FACTER_repopath}/src/bash/repo.sh
+
+function puppetrun() {
+	puppet apply ${FACTER_repopath}/${1} --environment $environment --detailed-exitcodes || if test $? != 2; then return $?; fi	
+	return $?
+}
+
+if ! dpkg -l | grep -q 'ii  puppetmaster'; then
+	apt-get install -y puppetmaster ruby
+fi
+puppetrun site/profile/manifests/master/provision.pp
